♻️ move buildCookieOptions into configuration.ts to break dependency cycle
