+++ packages/core/src/domain/configuration/configuration.spec.ts
@@ -5,7 +5,7 @@ import { DOCS_ORIGIN, MORE_DETAILS, display } from '../../tools/display'
 import { ExperimentalFeature, isExperimentalFeatureEnabled } from '../../tools/experimentalFeatures'
 import { TrackingConsent } from '../trackingConsent'
 import type { InitConfiguration } from './configuration'
-import { serializeConfiguration, validateAndBuildConfiguration } from './configuration'
+import { buildCookieOptions, serializeConfiguration, validateAndBuildConfiguration } from './configuration'
 
 describe('validateAndBuildConfiguration', () => {
   const clientToken = 'some_client_token'
@@ -208,6 +208,35 @@ describe('validateAndBuildConfiguration', () => {
         serializeConfiguration(EXHAUSTIVE_INIT_CONFIGURATION)
 
       expect(serializedConfiguration).toEqual(SERIALIZED_EXHAUSTIVE_INIT_CONFIGURATION)
+    })
+  })
+})
+
+describe('buildCookieOptions', () => {
+  const clientToken = 'abc'
+
+  it('should not be secure nor crossSite by default', () => {
+    const cookieOptions = buildCookieOptions({ clientToken })
+    expect(cookieOptions).toEqual({ secure: false, crossSite: false, partitioned: false })
+  })
+
+  it('should be secure when `useSecureSessionCookie` is truthy', () => {
+    const cookieOptions = buildCookieOptions({ clientToken, useSecureSessionCookie: true })
+    expect(cookieOptions).toEqual({ secure: true, crossSite: false, partitioned: false })
+  })
+
+  it('should be secure, crossSite and partitioned when `usePartitionedCrossSiteSessionCookie` is truthy', () => {
+    const cookieOptions = buildCookieOptions({ clientToken, usePartitionedCrossSiteSessionCookie: true })
+    expect(cookieOptions).toEqual({ secure: true, crossSite: true, partitioned: true })
+  })
+
+  it('should have domain when `trackSessionAcrossSubdomains` is truthy', () => {
+    const cookieOptions = buildCookieOptions({ clientToken, trackSessionAcrossSubdomains: true })
+    expect(cookieOptions).toEqual({
+      secure: false,
+      crossSite: false,
+      partitioned: false,
+      domain: jasmine.any(String),
     })
   })
 })
