+++ packages/core/src/domain/session/storeStrategies/sessionInCookie.spec.ts
@@ -3,9 +3,10 @@ import type { CookieStoreWindow } from '../../../browser/browser.types'
 import { Observable } from '../../../tools/observable'
 import type { SessionState } from '../sessionState'
 import type { Configuration, InitConfiguration } from '../../configuration'
+import { buildCookieOptions } from '../../configuration'
 import { SESSION_COOKIE_EXPIRATION_DELAY, SESSION_TIME_OUT_DELAY, SessionPersistence } from '../sessionConstants'
 import { CookieApi, LEGACY_SESSION_STORE_KEY } from './sessionStoreStrategy'
-import { buildCookieOptions, createCookieAccess, selectCookieStrategy, initCookieStrategy } from './sessionInCookie'
+import { createCookieAccess, selectCookieStrategy, initCookieStrategy } from './sessionInCookie'
 
 const DEFAULT_INIT_CONFIGURATION = { clientToken: 'abc', trackAnonymousUser: true }
 
@@ -227,35 +228,6 @@ describe('session in cookie strategy', () => {
       await strategy.setSessionState(() => ({ id: '123', created: '0' }), 'updateState')
 
       expect(mockCookie.getLastExpireDelay()).toBe(SESSION_TIME_OUT_DELAY)
-    })
-  })
-
-  describe('build cookie options', () => {
-    const clientToken = 'abc'
-
-    it('should not be secure nor crossSite by default', () => {
-      const cookieOptions = buildCookieOptions({ clientToken })
-      expect(cookieOptions).toEqual({ secure: false, crossSite: false, partitioned: false })
-    })
-
-    it('should be secure when `useSecureSessionCookie` is truthy', () => {
-      const cookieOptions = buildCookieOptions({ clientToken, useSecureSessionCookie: true })
-      expect(cookieOptions).toEqual({ secure: true, crossSite: false, partitioned: false })
-    })
-
-    it('should be secure, crossSite and partitioned when `usePartitionedCrossSiteSessionCookie` is truthy', () => {
-      const cookieOptions = buildCookieOptions({ clientToken, usePartitionedCrossSiteSessionCookie: true })
-      expect(cookieOptions).toEqual({ secure: true, crossSite: true, partitioned: true })
-    })
-
-    it('should have domain when `trackSessionAcrossSubdomains` is truthy', () => {
-      const cookieOptions = buildCookieOptions({ clientToken, trackSessionAcrossSubdomains: true })
-      expect(cookieOptions).toEqual({
-        secure: false,
-        crossSite: false,
-        partitioned: false,
-        domain: jasmine.any(String),
-      })
     })
   })
