+++ packages/core/src/domain/session/storeStrategies/sessionInCookie.ts
@@ -1,7 +1,7 @@
 import { isEmptyObject } from '../../../tools/utils/objectUtils'
 import type { CookieOptions } from '../../../browser/cookie'
-import { getCurrentSite, getCookies } from '../../../browser/cookie'
-import type { Configuration, InitConfiguration } from '../../configuration'
+import { getCookies } from '../../../browser/cookie'
+import type { Configuration } from '../../configuration'
 import { SESSION_COOKIE_EXPIRATION_DELAY, SESSION_TIME_OUT_DELAY, SessionPersistence } from '../sessionConstants'
 import type { SessionState } from '../sessionState'
 import { toSessionString, toSessionState } from '../sessionState'
@@ -152,25 +152,6 @@ function buildSessionString(sessionState: SessionState, cookieOptions: CookieOpt
   return toSessionString(
     isEmptyObject(sessionState) ? sessionState : { ...sessionState, c: encodeCookieOptions(cookieOptions) }
   )
-}
-
-export function buildCookieOptions(initConfiguration: InitConfiguration): CookieOptions | undefined {
-  const cookieOptions: CookieOptions = {}
-
-  cookieOptions.secure =
-    !!initConfiguration.useSecureSessionCookie || !!initConfiguration.usePartitionedCrossSiteSessionCookie
-  cookieOptions.crossSite = !!initConfiguration.usePartitionedCrossSiteSessionCookie
-  cookieOptions.partitioned = !!initConfiguration.usePartitionedCrossSiteSessionCookie
-
-  if (initConfiguration.trackSessionAcrossSubdomains) {
-    const currentSite = getCurrentSite()
-    if (!currentSite) {
-      return
-    }
-    cookieOptions.domain = currentSite
-  }
-
-  return cookieOptions
 }
 
 function encodeCookieOptions(cookieOptions: CookieOptions): string {
